Build 1.1.103 Исправлена некорректная индикация состоянии сессии Заряд батареи будет передаваться как int

+++ iOsmo/TcpConnection.swift
@@ -179,7 +179,7 @@ open class TcpConnection: BaseTcpConnection {
     
     open func sendBatteryStatus(){
         UIDevice.current.isBatteryMonitoringEnabled = true
-        let level = UIDevice.current.batteryLevel * 100
+        let level = Int(UIDevice.current.batteryLevel * 100)
         var state = 0;
         if (UIDevice.current.batteryState == .charging) {
             state = 1;
